+++ src/generic.rs
@@ -1,11 +1,9 @@
 use super::bits::BitAggregation;
+use super::matrix::Matrix;
 use bitcoin::{bip32::Xpriv, NetworkKind};
 use serde::Serialize;
 use std::fmt::Debug;
 use thiserror::Error;
-
-/// type alias for any diagram
-pub type Matrix<const H: usize, const W: usize, T> = [[Option<T>; W]; H];
 
 /// Generic Diagram  
 ///   diagram implementation for any matrix
@@ -91,38 +89,10 @@ pub enum GenericError {
 /// GenericResult
 pub type GenericResult<T = ()> = Result<T, GenericError>;
 
-/// transform to generic diagram
-pub trait ToMatrix<T> {
-    /// transform to matrix
-    fn to_matrix<const H: usize, const W: usize>(self) -> Matrix<H, W, T>;
-}
-
-impl<T> ToMatrix<T> for Vec<Option<T>> {
-    fn to_matrix<const H: usize, const W: usize>(mut self) -> Matrix<H, W, T> {
-        self.resize_with(H * W, || None);
-        self.reverse();
-        core::array::from_fn(|_| core::array::from_fn(|_| self.pop().unwrap()))
-    }
-}
-
-impl<T: std::fmt::Debug> ToMatrix<T> for Vec<Vec<Option<T>>> {
-    fn to_matrix<const H: usize, const W: usize>(mut self) -> Matrix<H, W, T> {
-        self.resize_with(H, || [const { None }; W].into_iter().collect());
-        self.into_iter()
-            .map(|mut r| {
-                r.resize_with(W, || None);
-                r.try_into().unwrap()
-            })
-            .take(H)
-            .collect::<Vec<[Option<T>; W]>>()
-            .try_into()
-            .unwrap()
-    }
-}
-
 #[cfg(test)]
 mod generic_test {
     use super::*;
+    use crate::ToMatrix;
     use bitcoin::hex::DisplayHex;
 
     #[test]
